+++ CHANGES.rst
@@ -20,6 +20,10 @@ New features:
 
 Bug fixes:
 
+- Fix controlpanel quickinstaller view: 
+  A not yet installed product must not return any upgrade info.
+  [jensens]
+
 - Fix to make plone/plone.session#11 work:
   Make test for installation of  plone.session more explicit.
   [jensens]